OPENMEETINGS-722 minor changes.
git-svn-id: https://svn.apache.org/repos/asf/openmeetings/trunk/singlewebapp@1508814 13f79535-47bb-0310-9956-ffa450edef68

+++ src/org/apache/openmeetings/web/app/WebSession.java
@@ -248,13 +248,13 @@ public class WebSession extends AbstractAuthenticatedWebSession {
 		return states.get(0);
 	}
 
-	public OmTimeZone getOmTimeZoneByBrowserLocale(int offset){
+	public OmTimeZone getOmTimeZoneByBrowserLocale(double offset){
 		TimeZone tz = Calendar.getInstance(getBrowserLocale()).getTimeZone();
 		OmTimeZone omTZ = getBean(OmTimeZoneDao.class).getOmTimeZoneByIcal(tz.getID());
 		if (omTZ == null){
 			List<OmTimeZone> omTimeZones = getBean(OmTimeZoneDao.class).getOmTimeZones();
 			for (OmTimeZone timeZone : omTimeZones){
-				long tzOffset = TimeZone.getTimeZone(timeZone.getIcal()).getRawOffset() / 3600000;
+				double tzOffset = TimeZone.getTimeZone(timeZone.getIcal()).getRawOffset() / 3600000;
 				if (tzOffset ==  offset){
 					return timeZone;  
 				}
